requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
langchain>=0.1.0
langchain-text-splitters>=0.0.1
langchain-google-genai>=1.0.0
langchain-core>=0.1.0
faiss-cpu>=1.7.4
google-generativeai>=0.3.0
gitpython>=3.1.40
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=1.24.0,<2.0.0
